Refactor code and implemented relations between User(Information) and UserDetail(Sign/signup)

+++ src/main/java/com/samucatezu/pimbackend/Services/InsuranceService.java
@@ -46,7 +46,7 @@ public class InsuranceService {
     }
 
     // update a product
-    public void updateInsurance(Integer insuranceID, InsuranceDTO insuranceDTO, Category category) {
+    public void updateInsurance(Long insuranceID, InsuranceDTO insuranceDTO, Category category) {
         Insurance insurance = getInsuranceFromDto(insuranceDTO, category);
         // set the id for updating
         insurance.setId(insuranceID);
